+++ bean-searcher-doc/docs/.vitepress/theme/index.ts
@@ -2,12 +2,14 @@
 import type { Theme } from 'vitepress'
 import DefaultTheme from 'vitepress/theme'
 import MyLayout from './MyLayout.vue'
+import ImagePreview from './ImagePreview.vue'
 import './style.css'
 
 export default {
   extends: DefaultTheme,
   Layout: MyLayout,
   enhanceApp({ app, router, siteData }) {
-    // ...
+    // 注册为全局组件
+    app.component('ImagePreview', ImagePreview)
   }
 } satisfies Theme