+++ bean-searcher/src/main/java/cn/zhxu/bs/param/Paging.java
@@ -1,5 +1,7 @@
 package cn.zhxu.bs.param;
 
+import cn.zhxu.bs.implement.DefaultSqlResolver;
+
 /**
  * 分页参数
  * @author Troy.Zhou @ 2021-10-31
@@ -16,6 +18,12 @@ public class Paging {
      * 分页：查询偏移条数（用于分页）
      */
     private long offset;
+
+    /**
+     * 是否有排序字段（内部字段，由 {@link DefaultSqlResolver } 自动管理）
+     * @since v4.8.3
+     */
+    private boolean ordering;
 
     public Paging(int size, long offset) {
         this.size = size;
@@ -36,6 +44,14 @@ public class Paging {
 
     public void setOffset(long offset) {
         this.offset = offset;
+    }
+
+    public boolean isOrdering() {
+        return ordering;
+    }
+
+    public void setOrdering(boolean ordering) {
+        this.ordering = ordering;
     }
 
 }